Moved Player model to models package added getPlayerByUsernamemethod to DatabaseHandler class at server
