+++ Client/src/client/Client.java
@@ -9,15 +9,7 @@ import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
-import screens.AfterGameScreenUI;
-import screens.ClientMainScreenUI;
-import screens.GameScreenUI;
-import screens.LocalPlayersNamesUI;
-import screens.LoginRegisterScreenUI;
 import screens.Navigation;
-import screens.OnlineUsersUI;
-import screens.PlayerProfileScreenUI;
-import screens.RegisterScreenUI;
 
 public class Client extends Application {
 
@@ -27,74 +19,7 @@ public class Client extends Application {
     public void start(Stage stage) throws Exception {
         
         Navigation navigation = new Navigation(stage);
-        
-        
-        //ClientMainScreenUI clientMainScreenLabel = new ClientMainScreenUI();
-        //Scene mainScreenScene = new Scene(clientMainScreenLabel);
-        /*
-        ClientMainScreenUI clientMainScreenLabel = new ClientMainScreenUI();
-        Scene mainScreenScene = new Scene(clientMainScreenLabel);
-
-        PlayerProfileScreenUI palyerProfileScreenLabel = new PlayerProfileScreenUI();
-        Scene palyerProfileScreenScene = new Scene(palyerProfileScreenLabel);
-
-        GameScreenUI gameScreenLabel = new GameScreenUI();
-        Scene gameScreenScene = new Scene(gameScreenLabel);
-
-        LocalPlayersNamesUI localPlayersNamesLabel = new LocalPlayersNamesUI();
-        Scene localPlayersNamesScene = new Scene(localPlayersNamesLabel);
-
-        LoginRegisterScreenUI loginRegisterScreenLabel = new LoginRegisterScreenUI();
-        Scene loginRegisterScreenScene = new Scene(loginRegisterScreenLabel);
-
-        RegisterScreenUI registerScreenLabel = new RegisterScreenUI();
-        Scene registerScreenScene = new Scene(registerScreenLabel);
-
-        OnlineUsersUI onlineUsersLabel = new OnlineUsersUI();
-        Scene onlineUsersScene = new Scene(onlineUsersLabel);
-
-        AfterGameScreenUI afterGameScreenLabel = new AfterGameScreenUI();
-        Scene afterGameScreenScene = new Scene(afterGameScreenLabel);
-
-        clientMainScreenLabel.getProfileBtn().setOnAction(event -> {
-            stage.setScene(palyerProfileScreenScene);
-        });
-        clientMainScreenLabel.getComputerBtn().setOnAction(event -> {
-
-        });
-        clientMainScreenLabel.getLocalBtn().setOnAction(event -> {
-            stage.setScene(localPlayersNamesScene);
-        });
-        clientMainScreenLabel.getOnlineBtn().setOnAction(event -> {
-            stage.setScene(loginRegisterScreenScene);
-        });
-        clientMainScreenLabel.getMainExitBtn().setOnAction(event -> {
-            stage.setScene(afterGameScreenScene);
-        });
-        localPlayersNamesLabel.getBtnStart().setOnAction(event -> {
-            stage.setScene(gameScreenScene);
-        });
-        localPlayersNamesLabel.getBtnCancel().setOnAction(event -> {
-            stage.setScene(mainScreenScene);
-        });
-
-//        loginRegisterScreenLabel.getLoginBtn().setOnAction(event -> {
-//            
-//        });
-
-        gameScreenLabel.getExitGameBtn().setOnAction(event -> {
-            stage.setScene(mainScreenScene);
-        });
-*/
-//        mainScreenScene.getStylesheets().add(getClass().getResource("style.css").toString());
-//        mainScreenScene.getStylesheets().add(getClass().getResource("/resources/buttons.css").toString());
-//        mainScreenScene.getStylesheets().add(getClass().getResource("/resources/background.css").toString());
-//        mainScreenScene.getStylesheets().add(getClass().getResource("/resources/labels.css").toString());
-//        mainScreenScene.getStylesheets().add(getClass().getResource("/resources/regbg.css").toString());
-//        mainScreenScene.getStylesheets().add(getClass().getResource("/resources/transparentButton.css").toString());
-        
         navigation.goTo("/screens/ClientMainScreen.fxml");
-        //stage.setScene(mainScreenScene);
         stage.setResizable(false);
         stage.show();
